requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
pyarrow>=10.0.0
pytest>=7.0.0
